feat: product service refactoring
